login and signup back end codes are ready please check the comments and uncomment if there is no conflict
