+++ db/db.js
@@ -1,0 +1,15 @@
+// var mysql=require('mysql');
+//  var connection=mysql.createConnection({
+//    host:'localhost',
+//    user:'root',
+//    password:'your password',
+//    database:'myapp'
+//  });
+// connection.connect(function(error){
+//    if(!!error){
+//      console.log(error);
+//    }else{
+//      console.log('Connected!:)');
+//    }
+//  });  
+// module.exports = connection;
