+++ public/src/login.html
@@ -52,7 +52,7 @@
             </div>
         </div>
     </div>
-    <script src="../app.js"></script>
+    <script src="./login.js"></script>
 </body>
 
 </html>
