+++ public/src/signup.html
@@ -54,7 +54,7 @@
             </div>
         </div>
     </div>
-    <script src="../app.js"></script>
+    <script src="./signUp.js"></script>
 </body>
 
 </html>
